requirements.txt
streamlit
pandas
numpy
scikit-learn
shap
matplotlib
pydantic
python-dotenv
reportlab
openpyxl
rapidfuzz
Unidecode
plotly